requirements.txt
# requirements.txt
# Credit Fraud Platform — Python dependencies

# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
python-multipart==0.0.9

# ─── Database ─────────────────────────────────────────────────────────────────
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.28

# ─── Ray ─────────────────────────────────────────────────────────────────────
ray[data,serve]==2.9.3

# ─── Graph ───────────────────────────────────────────────────────────────────
neo4j==5.17.0
gremlinpython==3.7.1          # AWS Neptune

# ─── ML ──────────────────────────────────────────────────────────────────────
scikit-learn==1.4.1
xgboost==2.0.3
lightgbm==4.3.0
numpy==1.26.4
pandas==2.2.1
joblib==1.3.2

# ─── LLM ─────────────────────────────────────────────────────────────────────
openai==1.13.3

# ─── AWS ─────────────────────────────────────────────────────────────────────
boto3==1.34.51
botocore==1.34.51

# ─── Kafka ───────────────────────────────────────────────────────────────────
confluent-kafka==2.3.0

# ─── Utilities ────────────────────────────────────────────────────────────────
httpx==0.27.0
python-dotenv==1.0.1
structlog==24.1.0
prometheus-client==0.20.0
tenacity==8.2.3               # retry logic for resilience

# ─── Testing ─────────────────────────────────────────────────────────────────
pytest==8.1.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.27.0                 # TestClient for FastAPI
